Moving commands to commands.taskqueue.go file
