+++ temporalcli/commands.taskqueue_test.go
@@ -4,7 +4,10 @@ import (
 	"encoding/json"
 	"time"
 
+	"github.com/google/uuid"
+	"github.com/temporalio/cli/temporalcli"
 	"go.temporal.io/api/enums/v1"
+	"go.temporal.io/api/workflowservice/v1"
 )
 
 func (s *SharedServerSuite) TestTaskQueue_Describe_Simple() {
@@ -58,4 +61,44 @@ func (s *SharedServerSuite) TestTaskQueue_Describe_Simple() {
 	s.NoError(res.Err)
 	s.NoError(json.Unmarshal(res.Stdout.Bytes(), &jsonOut))
 	s.GreaterOrEqual(10, len(jsonOut.TaskQueues))
+}
+
+func (s *SharedServerSuite) TestTaskQueue_ListPartition() {
+	testTaskQueue := uuid.NewString()
+	res := s.Execute(
+		"task-queue", "list-partition",
+		"--address", s.Address(),
+		"--task-queue", testTaskQueue,
+	)
+	s.Contains(res.Stdout.String(), testTaskQueue)
+	s.Contains(res.Stdout.String(), "Workflow Task Queue Partitions")
+	s.Contains(res.Stdout.String(), "Activity Task Queue Partitions")
+	s.NoError(res.Err)
+}
+
+func (s *SharedServerSuite) TestTaskQueue_ListPartitionInvalidNamespace() {
+	testTaskQueue := uuid.NewString()
+	res := s.Execute(
+		"task-queue", "list-partition",
+		"--address", s.Address(),
+		"--task-queue", testTaskQueue,
+		"--namespace", "invalid_namespace",
+	)
+	s.Error(res.Err)
+	s.Contains(res.Err.Error(), "Namespace invalid_namespace is not found")
+}
+
+func (s *SharedServerSuite) TestTaskQueue_ListPartitionJsonOutput() {
+	testTaskQueue := uuid.NewString()
+	res := s.Execute(
+		"task-queue", "list-partition",
+		"--address", s.Address(),
+		"--task-queue", testTaskQueue,
+		"--output", "json",
+	)
+	s.NoError(res.Err)
+	var listResp workflowservice.ListTaskQueuePartitionsResponse
+	s.NoError(temporalcli.UnmarshalProtoJSONWithOptions(res.Stdout.Bytes(), &listResp, true))
+	s.NotEmpty(listResp.ActivityTaskQueuePartitions)
+	s.NotEmpty(listResp.WorkflowTaskQueuePartitions)
 }
